+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 1.2.0: January 13, 2020
+- Improved autocomplete suggestions
+- Fixed snippets not working with other Intellisense extensions
+
 ### 1.1.0: January 10, 2020
 - Added formspec string syntax highlighting